% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{resetTimer}
\alias{resetTimer}
\title{Reset shinyTimer}
\usage{
resetTimer(session, inputId, hours = 0, minutes = 0, seconds = 0)
}
\arguments{
\item{session}{The session object from the shiny server function.}

\item{inputId}{The input ID corresponding to the UI element.}

\item{hours}{The new reset time in hours.}

\item{minutes}{The new reset time in minutes.}

\item{seconds}{The new reset time in seconds.}
}
\description{
Reset shinyTimer
}
