% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{runTimer}
\alias{runTimer}
\title{Countdown Timer Server Logic}
\usage{
runTimer(input, output, session, inputId, startTime)
}
\arguments{
\item{input}{The input object from the shiny server function.}

\item{output}{The output object from the shiny server function.}

\item{session}{The session object from the shiny server function.}

\item{inputId}{The input ID corresponding to the UI element.}

\item{startTime}{The starting time for the countdown.}
}
\description{
Countdown Timer Server Logic
}
