% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{stopTimer}
\alias{stopTimer}
\title{Stop \code{shinyTimer}}
\usage{
stopTimer(session, inputId)
}
\arguments{
\item{session}{The session object from the shiny server function.}

\item{inputId}{The input ID corresponding to the UI element.}
}
\description{
Stop \code{shinyTimer}
}
