% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countdown.R
\name{countdown}
\alias{countdown}
\title{A timer widget for RShiny}
\usage{
countdown(
  count = 5 * 60,
  format = "ms",
  sep = ":",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{count}{Countdown in seconds.}

\item{format}{A string indicating the displayed format (see details).}

\item{sep}{A string indicating the used separator. Default is ":".}

\item{width}{The width of the input, e.g. "400px", or "100%".}

\item{height}{The height of the input, e.g. "400px", or "100%".}

\item{elementId}{The input slot that will be used to access the value.}
}
\description{
The function creates a timer that counts down.
}
\details{
The format can contain "d" days, "h" hours, "m" minutes and or "s"
seconds. For example "ms" will only display minutes and seconds. If the next
higher unit is not given, the amount will be added to the last unit, e.g., 70
seconds instead of 1:10.
}
