% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{countDown}
\alias{countDown}
\title{Set shinyTimer in motion: count down}
\usage{
countDown(inputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{The input ID corresponding to the UI element.}

\item{session}{The session object from the shiny server function.}
}
\description{
Set shinyTimer in motion: count down
}
