% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{countDown}
\alias{countDown}
\title{Countdown Timer Server Logic}
\usage{
countDown(session, inputId)
}
\arguments{
\item{session}{The session object from the shiny server function.}

\item{inputId}{The input ID corresponding to the UI element.}
}
\description{
Countdown Timer Server Logic
}
