% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{countUp}
\alias{countUp}
\title{Set shinyTimer in motion: count up}
\usage{
countUp(session, inputId)
}
\arguments{
\item{session}{The session object from the shiny server function.}

\item{inputId}{The input ID corresponding to the UI element.}
}
\description{
Set shinyTimer in motion: count up
}
