% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{shinyTimer}
\alias{shinyTimer}
\title{Countdown Timer UI Component}
\usage{
shinyTimer(inputId, label = NULL, startTime = 10)
}
\arguments{
\item{inputId}{The input id.}

\item{label}{The label to display above the countdown.}

\item{startTime}{The starting time for the countdown.}
}
\value{
A shiny UI component for the countdown timer.
}
\description{
Countdown Timer UI Component
}
