% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{shinyTimer}
\alias{shinyTimer}
\title{Countdown Timer UI Component}
\usage{
shinyTimer(
  inputId,
  label = NULL,
  minutes = 0,
  seconds = 0,
  format = "simple",
  ...
)
}
\arguments{
\item{inputId}{The input id.}

\item{label}{The label to display above the countdown.}

\item{minutes}{An integer, the starting time in minutes for the countdown.}

\item{seconds}{An integer, the starting time in seconds for the countdown.}

\item{format}{The format of the countdown timer display (either "simple" or "clock").}

\item{...}{Any additional parameters you want to pass to the placeholder for the timer (\code{htmltools::tags$div}).}
}
\value{
A shiny UI component for the countdown timer.
}
\description{
Countdown Timer UI Component
}
