% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{updateShinyTimer}
\alias{updateShinyTimer}
\title{Update Countdown Timer}
\usage{
updateShinyTimer(session, inputId, label = NULL, start = NULL, format = NULL)
}
\arguments{
\item{session}{The session object from the shiny server function.}

\item{inputId}{The input ID corresponding to the UI element.}

\item{label}{The new label to be displayed above the countdown timer.}

\item{start}{The new starting time for the countdown.}

\item{format}{The new format of the countdown timer display (either "simple" or "clock").}
}
\description{
Update Countdown Timer
}
