% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{updateShinyTimer}
\alias{updateShinyTimer}
\title{Update shinyTimer widget}
\usage{
updateShinyTimer(
  session,
  inputId,
  hours = NULL,
  minutes = NULL,
  seconds = NULL,
  type = NULL,
  label = NULL,
  background = NULL
)
}
\arguments{
\item{session}{The session object from the shiny server function.}

\item{inputId}{The input ID corresponding to the UI element.}

\item{hours}{The new starting time in hours for the countdown.}

\item{minutes}{The new starting time in minutes for the countdown.}

\item{seconds}{The new starting time in seconds for the countdown.}

\item{type}{The new type of the countdown timer display ("simple", "mm:ss", "hh:mm:ss", "mm:ss.cs").}

\item{label}{The new label to be displayed above the countdown timer.}

\item{background}{The new shape of the timer's container ("none", "circle", "rectangle").}
}
\description{
Update shinyTimer widget
}
