% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{updateShinyTimer}
\alias{updateShinyTimer}
\title{Update \code{shinyTimer} widget}
\usage{
updateShinyTimer(
  session,
  inputId,
  minutes = NULL,
  seconds = NULL,
  format = NULL,
  label = NULL
)
}
\arguments{
\item{session}{The session object from the shiny server function.}

\item{inputId}{The input ID corresponding to the UI element.}

\item{minutes}{The new starting time in minutes for the countdown.}

\item{seconds}{The new starting time in seconds for the countdown.}

\item{format}{The new format of the countdown timer display ("simple", "clock", or "stopwatch").}

\item{label}{The new label to be displayed above the countdown timer.}
}
\description{
Update \code{shinyTimer} widget
}
